pyproject.toml
[tool.poetry]
name = "passexport"
version = "0.1.0"
description = ""
authors = ["Martijn Hemeryck <martijn.hemeryck@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
python-gnupg = "^0.4.6"
gitpython = "^3.1.9"

[tool.poetry.dev-dependencies]
ipython = "^7.18.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
